Reuse inputRefs utils' functions
