+++ front-end-mobile/src/features/authentication/utils/yupValidation.ts
